% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSGMM.R
\name{calculateBIC}
\alias{calculateBIC}
\title{Calculate the Bayesian Information Criterion (BIC)}
\usage{
calculateBIC(logL, params)
}
\arguments{
\item{logL}{Log-likelihood value of fitted model}

\item{params}{MSGMM model parameter output}
}
\value{
BIC score
}
\description{
Calculate the Bayesian Information Criterion (BIC)
}
\seealso{
\code{\link{getLoglikelihood}}
}
