% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculateJeffrey}
\alias{calculateJeffrey}
\title{Calculate Jeffreys divergence between two components}
\usage{
calculateJeffrey(mu1, covMat1, mu2, covMat2)
}
\description{
Calculate Jeffreys divergence between two components
}
