% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculateKL}
\alias{calculateKL}
\title{Calculate Kullback–Leibler (KL) divergence between two components}
\usage{
calculateKL(mu1, covMat1, mu2, covMat2)
}
\arguments{
\item{mu1}{Vector of means per dimension for first component}

\item{covMat1}{Covariance matrix of first component}

\item{mu2}{Vector of means per dimension for second component}

\item{covMat2}{Covariance matrix of second component}
}
\value{
Vector with log-likelihood values
}
\description{
Calculate Kullback–Leibler (KL) divergence between two components
}
