% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compDist}
\alias{compDist}
\title{Calculate distance matrix between components based on Jeffreys divergence}
\usage{
compDist(params)
}
\arguments{
\item{params}{MSGMM model parameter output}
}
\value{
Distance matrix
}
\description{
Calculate distance matrix between components based on Jeffreys divergence
}
