% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getEllipsoids}
\alias{getEllipsoids}
\title{Get component ellipsoids based on model parameters}
\usage{
getEllipsoids(params, dim1, dim2, conf = 0.95)
}
\arguments{
\item{params}{MSGMM model parameter output}

\item{dim1}{First dimension}

\item{dim2}{Second dimension}

\item{conf}{Confidence level}
}
\value{
Ellipsoid dimensions (x, y)
}
\description{
Get component ellipsoids based on model parameters
}
