% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSGMM.R
\name{getFeatures}
\alias{getFeatures}
\title{Get features from each file based on model parameters}
\usage{
getFeatures(files, params, type = "percentages")
}
\arguments{
\item{files}{Character vector with CSV or FCS filenames}

\item{params}{MSGMM model parameter output}

\item{type}{Type of features to extract ("counts" or "percentages")}
}
\value{
Matrix with features per file
}
\description{
Get features from each file based on model parameters
}
