% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSGMM.R
\name{getLoglikelihood}
\alias{getLoglikelihood}
\title{Calculate log-likelihood}
\usage{
getLoglikelihood(files, params)
}
\arguments{
\item{files}{Character vector with CSV or FCS filenames}

\item{params}{MSGMM model parameter output}
}
\value{
Log-likelihood values
}
\description{
Calculate log-likelihood
}
\seealso{
\code{\link{getLoglikelihoodValues}}
}
