% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSGMM.R
\name{getLoglikelihoodValues}
\alias{getLoglikelihoodValues}
\title{Calculate log-likelihood values}
\usage{
getLoglikelihoodValues(files, usecols = NULL, params)
}
\arguments{
\item{files}{Character vector with CSV filenames}

\item{usecols}{Vector specifying the variables (columns) to use}

\item{params}{MSGMM model parameter output}
}
\value{
Vector with log-likelihood values
}
\description{
Calculate log-likelihood values
}
\seealso{
\code{\link{getLoglikelihood}}
}
