% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSGMM.R
\name{predictLabels}
\alias{predictLabels}
\title{Assign cluster labels to new data}
\usage{
predictLabels(X, params)
}
\arguments{
\item{X}{Numeric matrix}

\item{params}{MSGMM model parameter output}
}
\value{
Cluster labels
}
\description{
Assign cluster labels to new data
}
