% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSGMM.R
\name{readData}
\alias{readData}
\title{Wrapper for reading files from different sources (CSV, FCS)}
\usage{
readData(file, usecols)
}
\arguments{
\item{file}{File path}

\item{usecols}{Vector specifying the variables (columns) to use}
}
\value{
Numeric matrix
}
\description{
Wrapper for reading files from different sources (CSV, FCS)
}
