% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{splitCovariances}
\alias{splitCovariances}
\title{Splits the concatenated covariance matrix into a list ordered by component}
\usage{
splitCovariances(params)
}
\description{
Splits the concatenated covariance matrix into a list ordered by component
}
