% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSGMM.R
\name{subsampleFiles}
\alias{subsampleFiles}
\title{Subsample data from a list of CSV or FCS files}
\usage{
subsampleFiles(init.files, usecols, init.size, seed)
}
\arguments{
\item{init.files}{Character vector CSV or FCS filenames to use for K-means}

\item{usecols}{Vector specifying the variables (columns) to use}

\item{init.size}{Number of data points to sample from each file for means initialization}

\item{seed}{Random seed for subsampling during means initialization}
}
\value{
Matrix with subsampled data
}
\description{
Subsample data from a list of CSV or FCS files
}
