% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSGMM.R
\name{MSGMM}
\alias{MSGMM}
\title{Fit multi-sample Gaussian Mixture Model (MSGMM)}
\usage{
MSGMM(
  files,
  K,
  usecols,
  init.means = NULL,
  init.files = NULL,
  init.size = 10000,
  seed = NULL,
  tol = 0.001,
  max.iter = 50,
  gamma = 1,
  lambda = 0.01,
  pooled = FALSE
)
}
\arguments{
\item{files}{Character vector with CSV or FCS filenames}

\item{K}{Integer specifying number of Gaussian components}

\item{usecols}{Numeric or character vector specifying the columns to use}

\item{init.means}{Initial values of means. (K-means used if not specified)}

\item{init.files}{Character vector with CSV or FCS filenames to use for K-means}

\item{init.size}{Number of data points to sample from each file for means initialization}

\item{seed}{Random seed for subsampling during means initialization}

\item{tol}{Convergence threshold for stopping criterion}

\item{max.iter}{Maximum number of EM iterations}

\item{gamma}{Starting value of component variance matrices}

\item{lambda}{Regularization parameter for component covariance estimation}

\item{pooled}{Logical flag to enable/disable multi-sample EM algorithm}
}
\value{
List of model parameters ("weights", "means", and "covariances")
}
\description{
Fit multi-sample Gaussian Mixture Model (MSGMM)
}
